test(SwipeCell): update test cases
