+++ src/action-bar/README.zh-CN.md
@@ -113,7 +113,7 @@ export default {
 | --- | --- | --- | --- |
 | text | 按钮文字 | _string_ | - |
 | type | 按钮类型，可选值为 `primary` `info` `warning` `danger` | _string_ | `default` |
-| color | 按钮颜色，支持传入`linear-gradient`渐变色 | _string_ | - |
+| color | 按钮颜色，支持传入 `linear-gradient` 渐变色 | _string_ | - |
 | icon `v2.4.4` | 左侧[图标名称](#/zh-CN/icon)或图片链接 | _string_ | - |
 | disabled | 是否禁用按钮 | _boolean_ | `false` | - |
 | loading | 是否显示为加载状态 | _boolean_ | `false` | - |
